Move base sensor receiver & worker to common module
