+++ app/src/full/java/io/homeassistant/companion/android/sensors/GeocodeSensorManager.kt
@@ -8,6 +8,7 @@ import android.os.Build
 import android.util.Log
 import com.google.android.gms.location.LocationServices
 import io.homeassistant.companion.android.R
+import io.homeassistant.companion.android.common.sensors.SensorManager
 import io.homeassistant.companion.android.database.AppDatabase
 import io.homeassistant.companion.android.database.sensor.Setting
 import io.homeassistant.companion.android.location.HighAccuracyLocationService
