+++ app/src/main/java/io/homeassistant/companion/android/sensors/AudioSensorManager.kt
@@ -5,6 +5,7 @@ import android.media.AudioDeviceInfo
 import android.media.AudioManager
 import android.os.Build
 import io.homeassistant.companion.android.R
+import io.homeassistant.companion.android.common.sensors.SensorManager
 
 class AudioSensorManager : SensorManager {
     companion object {
