+++ app/src/main/java/io/homeassistant/companion/android/sensors/KeyguardSensorManager.kt
@@ -5,6 +5,7 @@ import android.content.Context
 import android.os.Build
 import androidx.annotation.RequiresApi
 import io.homeassistant.companion.android.R
+import io.homeassistant.companion.android.common.sensors.SensorManager
 
 class KeyguardSensorManager : SensorManager {
     companion object {
