+++ app/src/main/java/io/homeassistant/companion/android/sensors/NextAlarmManager.kt
@@ -4,6 +4,7 @@ import android.app.AlarmManager
 import android.content.Context
 import android.util.Log
 import io.homeassistant.companion.android.R
+import io.homeassistant.companion.android.common.sensors.SensorManager
 import io.homeassistant.companion.android.database.AppDatabase
 import io.homeassistant.companion.android.database.sensor.Setting
 import java.text.SimpleDateFormat
