+++ app/src/main/java/io/homeassistant/companion/android/sensors/NotificationSensorManager.kt
@@ -11,6 +11,7 @@ import android.service.notification.StatusBarNotification
 import android.util.Log
 import androidx.core.app.NotificationManagerCompat
 import io.homeassistant.companion.android.R
+import io.homeassistant.companion.android.common.sensors.SensorManager
 
 class NotificationSensorManager : NotificationListenerService(), SensorManager {
     companion object {
