+++ app/src/main/java/io/homeassistant/companion/android/sensors/PhoneStateSensorManager.kt
@@ -7,6 +7,7 @@ import android.telephony.SubscriptionInfo
 import android.telephony.SubscriptionManager
 import android.telephony.TelephonyManager
 import io.homeassistant.companion.android.R
+import io.homeassistant.companion.android.common.sensors.SensorManager
 
 class PhoneStateSensorManager : SensorManager {
