+++ app/src/main/java/io/homeassistant/companion/android/sensors/SensorReceiver.kt
@@ -3,7 +3,6 @@ package io.homeassistant.companion.android.sensors
 import android.annotation.SuppressLint
 import android.app.NotificationManager
 import android.bluetooth.BluetoothAdapter
-import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.content.res.Configuration
@@ -14,17 +13,17 @@ import dagger.hilt.android.AndroidEntryPoint
 import io.homeassistant.companion.android.BuildConfig
 import io.homeassistant.companion.android.common.data.integration.IntegrationRepository
 import io.homeassistant.companion.android.common.data.integration.SensorRegistration
+import io.homeassistant.companion.android.common.sensors.BatterySensorManager
+import io.homeassistant.companion.android.common.sensors.LastUpdateManager
+import io.homeassistant.companion.android.common.sensors.SensorReceiverBase
 import io.homeassistant.companion.android.database.AppDatabase
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
 import java.util.Locale
-import javax.inject.Inject
 
 @AndroidEntryPoint
-class SensorReceiver : BroadcastReceiver() {
+class SensorReceiver : SensorReceiverBase() {
+
+    override val tag: String
+        get() = TAG
 
     companion object {
         const val TAG = "SensorReceiver"
@@ -59,21 +58,9 @@ class SensorReceiver : BroadcastReceiver() {
             "io.homeassistant.companion.android.background.REQUEST_SENSORS_UPDATE"
     }
 
-    private val ioScope: CoroutineScope = CoroutineScope(Dispatchers.IO + Job())
-
-    @Inject
-    lateinit var integrationUseCase: IntegrationRepository
-
-    private val chargingActions = listOf(
-        Intent.ACTION_BATTERY_LOW,
-        Intent.ACTION_BATTERY_OKAY,
-        Intent.ACTION_POWER_CONNECTED,
-        Intent.ACTION_POWER_DISCONNECTED
-    )
-
     // Suppress Lint because we only register for the receiver if the android version matches the intent
     @SuppressLint("InlinedApi")
-    private val skippableActions = mapOf(
+    override val skippableActions = mapOf(
         "android.app.action.NEXT_ALARM_CLOCK_CHANGED" to NextAlarmManager.nextAlarm.id,
         "android.bluetooth.device.action.ACL_CONNECTED" to BluetoothSensorManager.bluetoothConnection.id,
         "android.bluetooth.device.action.ACL_DISCONNECTED" to BluetoothSensorManager.bluetoothConnection.id,
@@ -88,60 +75,7 @@ class SensorReceiver : BroadcastReceiver() {
         AudioManager.RINGER_MODE_CHANGED_ACTION to AudioSensorManager.audioSensor.id
     )
 
-    override fun onReceive(context: Context, intent: Intent) {
-
-        if (skippableActions.containsKey(intent.action)) {
-            val sensor = skippableActions[intent.action]
-            if (!isSensorEnabled(context, sensor!!)) {
-                Log.d(
-                    TAG,
-                    String.format
-                    ("Sensor %s corresponding to received event %s is disabled, skipping sensors update", sensor, intent.action)
-                )
-                return
-            }
-        }
-
-        if (isSensorEnabled(context, LastUpdateManager.lastUpdate.id)) {
-            LastUpdateManager().sendLastUpdate(context, intent.action)
-            val sensorDao = AppDatabase.getInstance(context).sensorDao()
-            val allSettings = sensorDao.getSettings(LastUpdateManager.lastUpdate.id)
-            for (setting in allSettings) {
-                if (setting.value != "" && intent.action == setting.value) {
-                    val eventData = intent.extras?.keySet()?.map { it.toString() to intent.extras?.get(it).toString() }?.toMap()?.plus("intent" to intent.action.toString())
-                        ?: mapOf("intent" to intent.action.toString())
-                    Log.d(TAG, "Event data: $eventData")
-                    ioScope.launch {
-                        try {
-                            integrationUseCase.fireEvent(
-                                "android.intent_received",
-                                eventData as Map<String, Any>
-                            )
-                            Log.d(TAG, "Event successfully sent to Home Assistant")
-                        } catch (e: Exception) {
-                            Log.e(TAG, "Unable to send event data to Home Assistant", e)
-                        }
-                    }
-                }
-            }
-        }
-
-        ioScope.launch {
-            updateSensors(context, integrationUseCase)
-            if (chargingActions.contains(intent.action)) {
-                // Add a 5 second delay to perform another update so charging state updates completely.
-                // This is necessary as the system needs a few seconds to verify the charger.
-                delay(5000L)
-                updateSensors(context, integrationUseCase)
-            }
-        }
-    }
-
-    private fun isSensorEnabled(context: Context, id: String): Boolean {
-        return AppDatabase.getInstance(context).sensorDao().get(id)?.enabled == true
-    }
-
-    suspend fun updateSensors(
+    override suspend fun updateSensors(
         context: Context,
         integrationUseCase: IntegrationRepository
     ) {
