+++ common/src/main/java/io/homeassistant/companion/android/common/sensors/BatterySensorManager.kt
@@ -1,10 +1,10 @@
-package io.homeassistant.companion.android.sensors
+package io.homeassistant.companion.android.common.sensors
 
 import android.content.Context
 import android.content.Intent
 import android.content.IntentFilter
 import android.os.BatteryManager
-import io.homeassistant.companion.android.R
+import io.homeassistant.companion.android.common.R
 
 class BatterySensorManager : SensorManager {
 
@@ -16,38 +16,32 @@ class BatterySensorManager : SensorManager {
             R.string.basic_sensor_name_battery_level,
             R.string.sensor_description_battery_level,
             "battery",
-            "%",
-            stateClass = SensorManager.STATE_CLASS_MEASUREMENT,
-            entityCategory = SensorManager.ENTITY_CATEGORY_DIAGNOSTIC
+            "%"
         )
         private val batteryState = SensorManager.BasicSensor(
             "battery_state",
             "sensor",
             R.string.basic_sensor_name_battery_state,
-            R.string.sensor_description_battery_state,
-            entityCategory = SensorManager.ENTITY_CATEGORY_DIAGNOSTIC
+            R.string.sensor_description_battery_state
         )
         private val isChargingState = SensorManager.BasicSensor(
             "is_charging",
             "binary_sensor",
             R.string.basic_sensor_name_charging,
             R.string.sensor_description_charging,
-            "plug",
-            entityCategory = SensorManager.ENTITY_CATEGORY_DIAGNOSTIC
+            "plug"
         )
         private val chargerTypeState = SensorManager.BasicSensor(
             "charger_type",
             "sensor",
             R.string.basic_sensor_name_charger_type,
-            R.string.sensor_description_charger_type,
-            entityCategory = SensorManager.ENTITY_CATEGORY_DIAGNOSTIC
+            R.string.sensor_description_charger_type
         )
         private val batteryHealthState = SensorManager.BasicSensor(
             "battery_health",
             "sensor",
             R.string.basic_sensor_name_battery_health,
-            R.string.sensor_description_battery_health,
-            entityCategory = SensorManager.ENTITY_CATEGORY_DIAGNOSTIC
+            R.string.sensor_description_battery_health
         )
 
         private val batteryTemperature = SensorManager.BasicSensor(
@@ -56,9 +50,7 @@ class BatterySensorManager : SensorManager {
             R.string.basic_sensor_name_battery_temperature,
             R.string.sensor_description_battery_temperature,
             "temperature",
-            "°C",
-            stateClass = SensorManager.STATE_CLASS_MEASUREMENT,
-            entityCategory = SensorManager.ENTITY_CATEGORY_DIAGNOSTIC
+            "°C"
         )
     }
