+++ common/src/main/res/values/strings.xml
@@ -1,8 +1,25 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
+    <string name="basic_sensor_name_battery_health">Battery Health</string>
+    <string name="basic_sensor_name_battery_level">Battery Level</string>
+    <string name="basic_sensor_name_battery_state">Battery State</string>
+    <string name="basic_sensor_name_battery_temperature">Battery Temperature</string>
+    <string name="basic_sensor_name_charger_type">Charger Type</string>
+    <string name="basic_sensor_name_charging">Is Charging</string>
+    <string name="basic_sensor_name_last_update">Last Update Trigger</string>
     <string name="database_event_failure">Unable to send migration failure event to Home Assistant</string>
     <string name="database_migration_failed">Database migration failed, all sensor and widget data has been reset. Please re-enable your sensors and recreate your widgets.</string>
     <string name="sensor">Sensor</string>
     <string name="sensors">Sensors</string>
     <string name="sensor_description_none">No description</string>
+    <string name="updating_sensors">Updating Sensors</string>
+    <string name="sensor_description_battery_health">The health of the battery</string>
+    <string name="sensor_description_battery_level">The current battery level of the device</string>
+    <string name="sensor_description_battery_state">The current charging state of the battery</string>
+    <string name="sensor_description_battery_temperature">The current battery temperature</string>
+    <string name="sensor_description_charger_type">The type of charger plugged into the device currently</string>
+    <string name="sensor_description_charging">Whether or not the device is actively charging</string>
+    <string name="sensor_description_last_update">The intent for the last update that was sent, periodic updates will show as \"SensorWorker\". Enabling the \"Add New Intent\" toggle will create 1 setting to allow you to register for a broadcast intent. The toggle will switch back to off once a new setting is created so you will need to turn it back on to save more intents. You can also clear out the setting value to remove the setting in the next update. You must restart the application after making changes to these settings to take effect.</string>
+    <string name="sensor_name_battery">Battery Sensors</string>
+    <string name="sensor_name_last_update">Last Update Sensor</string>
 </resources>
