+++ wear/src/main/AndroidManifest.xml
@@ -18,6 +18,20 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.HomeAssistant"
         android:fullBackupContent="@xml/backup_rules">
+        <!-- Start things like SensorWorker on device boot -->
+        <receiver android:name=".sensors.SensorReceiver">
+            <intent-filter>
+                <action android:name="android.intent.action.BOOT_COMPLETED" />
+                <action android:name="android.intent.action.MY_PACKAGE_REPLACED" />
+                <action android:name="android.intent.action.MY_PACKAGE_SUSPENDED" />
+                <action android:name="android.intent.action.MY_PACKAGE_UNSUSPENDED" />
+                <action android:name="android.intent.action.QUICKBOOT_POWERON" />
+                <action android:name="com.htc.intent.action.QUICKBOOT_POWERON" />
+                <action android:name="android.app.action.NEXT_ALARM_CLOCK_CHANGED" />
+                <action android:name="android.bluetooth.device.action.ACL_CONNECTED" />
+                <action android:name="android.bluetooth.device.action.ACL_DISCONNECTED" />
+            </intent-filter>
+        </receiver>
         <uses-library
             android:name="com.google.android.wearable"
             android:required="true" />
